# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Bakebee Admin API
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /:
    get:
      summary: Root endpoint
      responses:
        "200":
          description: Basic greeting
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/summary:
    get:
      summary: Admin summary
      responses:
        "200":
          description: Summary data
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: integer
                  orders:
                    type: integer
  /categories:
    get:
      summary: List categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      summary: Get category by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Category not found
    put:
      summary: Update category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      summary: Deactivate category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Category deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}/status:
    patch:
      summary: Activate or deactivate category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCategoryActiveRequest"
      responses:
        "200":
          description: Updated category status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /sub-categories:
    get:
      summary: List sub-categories
      responses:
        "200":
          description: Sub-category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubCategory"
    post:
      summary: Create sub-category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubCategoryRequest"
      responses:
        "201":
          description: Created sub-category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubCategory"
  /sub-categories/{id}:
    get:
      summary: Get sub-category by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Sub-category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubCategory"
        "404":
          description: Sub-category not found
    put:
      summary: Update sub-category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubCategoryRequest"
      responses:
        "200":
          description: Updated sub-category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubCategory"
    delete:
      summary: Deactivate sub-category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Sub-category deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubCategory"
  /sub-categories/{id}/status:
    patch:
      summary: Activate or deactivate sub-category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSubCategoryActiveRequest"
      responses:
        "200":
          description: Updated sub-category status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubCategory"
components:
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category_name:
          type: string
        is_active:
          type: boolean
        sub_categories:
          type: array
          items:
            $ref: "#/components/schemas/SubCategory"
    SubCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sub_category_name:
          type: string
        main_category_id:
          type: string
          format: uuid
        is_active:
          type: boolean
    CreateCategoryRequest:
      type: object
      required:
        - category_name
      properties:
        category_name:
          type: string
          minLength: 2
    UpdateCategoryRequest:
      type: object
      properties:
        category_name:
          type: string
          minLength: 2
    SetCategoryActiveRequest:
      type: object
      required:
        - is_active
      properties:
        is_active:
          type: boolean
    CreateSubCategoryRequest:
      type: object
      required:
        - sub_category_name
        - main_category_id
      properties:
        sub_category_name:
          type: string
          minLength: 2
        main_category_id:
          type: string
          format: uuid
    UpdateSubCategoryRequest:
      type: object
      properties:
        sub_category_name:
          type: string
          minLength: 2
        main_category_id:
          type: string
          format: uuid
    SetSubCategoryActiveRequest:
      type: object
      required:
        - is_active
      properties:
        is_active:
          type: boolean